+++ app/src/main/java/ru/frogogo/whitelabel/data/model/api/receipt/Receipt.kt
@@ -3,6 +3,7 @@ package ru.frogogo.whitelabel.data.model.api.receipt
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
 import ru.frogogo.whitelabel.dictionary.ReceiptState
+import java.math.BigDecimal
 import java.util.Date
 
 @JsonClass(generateAdapter = true)
@@ -12,7 +13,7 @@ data class Receipt(
   @Json(name = "number")
   val number: Int,
   @Json(name = "sum")
-  val sum: Int,
+  val sum: BigDecimal,
   @Json(name = "state")
   val state: ReceiptState,
   @Json(name = "timestamp")
